Do not start up node service on hubs
The main selenium class works as a kind of node classifier. If the
current node is the hub (determined by comparing $::fqdn with the
$selenium::hub_host parameter), a hub service is started. Previously,
a node service was always started automatically, whether this host was
determined to be the hub or not. This behavior is actually undesirable
in most cases -- for the sake of isolation & stability, you probably
want the hub to run by itself without any browser processes spawning
on the same machine and sporadically affecting load. So I have changed
the default behavior to start up EITHER a hub OR a node, but not both.

To retain previous behavior (where a node service is started on hub,
too), you can modify your roles / node definitions so that

  include selenium

becomes

  include selenium
  include selenium::node